+++ src/main/java/org/usfirst/frc4904/robot/subsystems/arm/ArmExtensionSubsystem.java
@@ -4,27 +4,23 @@
 package org.usfirst.frc4904.robot.subsystems;
 
 import java.util.function.DoubleSupplier;
-import java.util.stream.DoubleStream.DoubleMapMultiConsumer;
 
 import org.opencv.core.Mat.Tuple2;
-import org.usfirst.frc4904.robot.RobotMap;
 import org.usfirst.frc4904.standard.custom.motioncontrollers.ezControl;
 import org.usfirst.frc4904.standard.custom.motioncontrollers.ezMotion;
 import org.usfirst.frc4904.standard.subsystems.motor.TalonMotorSubsystem;
 
 import edu.wpi.first.math.controller.ArmFeedforward;
 import edu.wpi.first.math.trajectory.TrapezoidProfile;
+import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 public class ArmExtensionSubsystem extends SubsystemBase {
     
     private final TalonMotorSubsystem motor;
     private final static double EXTENSION_SPEED = 0.5;
-    private final static double SPOOL_DIAMETER = 1; // inches
-    private final static double TOTAL_ARM_LENGTH = 71; //inches
-    private final static double RETRACTED_ARM_LENGTH = 31; //inches
+    private final static double SPOOL_DIAMETER = Units.inchesToMeters(1);
     public final static double SPOOL_CIRCUMFERENCE = Math.PI * SPOOL_DIAMETER; // Math.PI * SPOOL_DIAMETER
     private final static double GEARBOX_RATIO = 12; // 12:1 
     private final ArmFeedforward feedforward;
@@ -52,21 +48,21 @@ public class ArmExtensionSubsystem extends SubsystemBase {
     }
 
     public double getCurrentExtensionLength() {
-        return getExtensionLength(motor.getSensorPositionRotations());
+        return revsToExtensionLength(motor.getSensorPositionRotations());
     }
 
-    public double getExtensionLength(double rotations) {
+    public double revsToExtensionLength(double rotations) {
         final double number_of_spool_rotations = rotations/GEARBOX_RATIO;
         final double extensionLength = number_of_spool_rotations * SPOOL_CIRCUMFERENCE;
         return extensionLength;
     }
 
-    public Command c_holdExtension(double extensionLengthInches, double maxVelocity, double maxAcceleration) {
+    public Command c_holdExtension(double extensionLengthMeters, double maxVelocity, double maxAcceleration) {
         ezControl controller = new ezControl(1, 1, 1, 
                                             (double position, double velocity) -> this.feedforward.calculate(angleDealer.getAsDouble() + Math.PI/2, velocity, 0));
         
         TrapezoidProfile profile = new TrapezoidProfile(new TrapezoidProfile.Constraints(maxVelocity, maxAcceleration), 
-                                                        new TrapezoidProfile.State(extensionLengthInches, 0), 
+                                                        new TrapezoidProfile.State(extensionLengthMeters, 0), 
                                                         new TrapezoidProfile.State(getCurrentExtensionLength(), 0));
 
         return new ezMotion(controller, this::getCurrentExtensionLength, motor::setVoltage, (double t) -> new Tuple2<Double>(profile.calculate(t).position, profile.calculate(t).velocity), this);
